First rip on inheritance chain,  see desc
Okay so there is interface DB, then AbstractDB. Media extends it and just abstracts the REST methods. The actual implementation does live in the Movie/TVShow models,  but it feels like when doing this, the models are just being used as database interfaces. Find out what the right balance is between in-memory representations and the connection. The put method for movie is static, which compiles but feels weird.  No instantiation going on there. Maybe this is okay because we can just have seperate DB interfaces associated with each ADT? Then those interfaces are linked or something to a collection of in-memory elements? Or is it a bad idea to in-mem all of them since that defeats the purpose of a db? dunno